Classrooms

+++ app/Http/Controllers/Grades/GradeController.php
@@ -3,6 +3,8 @@
 
 namespace App\Http\Controllers\Grades;
 use App\Http\Controllers\Controller;
+use CodeZero\UniqueTranslation\UniqueTranslationRule;
+use App\Models\Classroom;
 
 use App\Http\Requests\StoreGrades;
 use App\Models\Grade;
@@ -32,24 +34,23 @@ class GradeController extends Controller
   public function store(StoreGrades $request)
   {
 
-    // if (Grades::Where('Name->ar',$request->Name_ar)->orWhere('Name->en',$request->Name)->exists())
-    // {
-    //   return  redirect()->back()->WithErrors(trans('Grades_trans.Processes'));
-    return redirect()->back()->withErrors(['error' => trans('Grades_trans.exists')]);
-    // }
+    //   if (Grade::where('Name->ar', $request->Name)->orWhere('Name->en',$request->Name_en)->exists()) {
+
+    //       return redirect()->back()->withErrors(trans('Grades_trans.exists'));
+    //   }
 
 
+ try {
 
-      try {
           $validated = $request->validated();
           $Grade = new Grade();
-
-        //   $translations = [
-        //       'en' => $request->Name_en,
-        //       'ar' => $request->Name
-        //   ];
-        //   $Grade->setTranslations('Name', $translations);
-
+          /*
+          $translations = [
+              'en' => $request->Name_en,
+              'ar' => $request->Name
+          ];
+          $Grade->setTranslations('Name', $translations);
+          */
           $Grade->Name = ['en' => $request->Name_en, 'ar' => $request->Name];
           $Grade->Notes = $request->Notes;
           $Grade->save();
@@ -72,8 +73,8 @@ class GradeController extends Controller
    */
    public function update(StoreGrades $request)
  {
-
    try {
+
        $validated = $request->validated();
        $Grades = Grade::findOrFail($request->id);
        $Grades->update([
@@ -83,7 +84,6 @@ class GradeController extends Controller
        toastr()->success(trans('messages.Update'));
        return redirect()->route('Grades.index');
    }
-
    catch
    (\Exception $e) {
        return redirect()->back()->withErrors(['error' => $e->getMessage()]);
@@ -99,12 +99,22 @@ class GradeController extends Controller
   public function destroy(Request $request)
   {
 
-    $Grades = Grade::findOrFail($request->id)->delete();
-    toastr()->error(trans('messages.Delete'));
-    return redirect()->route('Grades.index');
+    $MyClass_id = Classroom::where('Grade_id',$request->id)->pluck('Grade_id');
+
+      if($MyClass_id->count() == 0){
+
+          $Grades = Grade::findOrFail($request->id)->delete();
+          toastr()->error(trans('messages.Delete'));
+          return redirect()->route('Grades.index');
+      }
+
+      else{
+
+          toastr()->error(trans('Grades_trans.delete_Grade_Error'));
+          return redirect()->route('Grades.index');
+
+      }
 
   }
 
 }
-
-?>
